refactor: add ArtworkFilterOptionItem component
